Cardano.Timeseries.API

Documentation

data ExecutionError where

Constructors

ParsingErrorWhileExecuting

Fields

:: { message :: Text

} -> ExecutionError

ElabErrorWhileExecuting

Fields

:: { message :: Text

} -> ExecutionError

InterpErrorWhileExecuting

Fields

:: { message :: Text

} -> ExecutionError

execute :: Store s Double => s -> Config -> Timestamp -> Text -> Either ExecutionError Value

A pure interface to the composite of parsing, elaboration and interpretation of timeseries queries.

module Cardano.Timeseries.AsText

module Cardano.Timeseries.Elab.Expr.Parser

module Cardano.Timeseries.Interp

module Cardano.Timeseries.Interp.Config

module Cardano.Timeseries.Store.Flat.Parser

elab :: Expr -> ElabM Expr

Given a surface query Expr elaborate it into a regular "semantic" query Expr.

initialSt :: Set MetricIdentifier -> St

fromFlat :: Flat a -> Tree a

type MetricIdentifier = Text

Each series in the (metric) store can be identified by a metric name.

type Timestamp = Word64

Milliseconds since epoch.

newtype InterpError

Constructors

InterpError

Fields

message :: Text

type InterpM a = ExceptT InterpError (State Int) a

data Value where

A model of values that queries interpret into.

Constructors

Scalar :: Double -> Value
A scalar.

RangeVector :: TimeseriesVector Value -> Value
A range vector.

InstantVector :: InstantVector Value -> Value
An instant vector.

Pair :: Value -> Value -> Value
A pair.

Unit :: Value
Unit.

Truth :: Value
Truth.

Falsity :: Value
Falsity.

Duration :: Word64 -> Value
Duration (milliseconds)

Timestamp :: Word64 -> Value
Timestamp (milliseconds since epoch)

Function :: FunctionValue -> Value
Function

Nil :: Value

Cons :: Value -> Value -> Value

Text :: Text -> Value
Text

class Store s a | s -> a where

A type-class witnessing that s is a metric-store of a.

Methods

insert :: s -> MetricIdentifier -> Instant a -> s

Insert an instant into the store under a metric name.

truncate :: s -> Timestamp -> s

Delete all entries older than the specified Timestamp from the store

evaluate :: s -> MetricIdentifier -> Timestamp -> InstantVector a

Compute a point vector of type a such that the timestamp of every point in the vector | lies within the staleness window of the target timestamp and is the most recent of all | points in the store sharing a series.

earliest :: s -> MetricIdentifier -> Maybe Timestamp

Find the earliest occurrence of the metric in the store, if any.

latest :: s -> MetricIdentifier -> Maybe Timestamp

Find the latest occurrence of the metric in the store, if any.

new :: s

An empty store.

metrics :: s -> Set MetricIdentifier

count :: s -> Int

Total number of (metric, labels, timestamp, value) tuples.

type Flat a = [Point a]

type Tree a = Map MetricIdentifier (Map Timestamp [Point a])