Cardano.Timeseries.Component

Documentation

data TimeseriesConfig

Constructors

TimeseriesConfig

Fields

retentionMillis :: Word64

How long the store entries are retained for (ms).

pruningPeriodSec :: Maybe Word64

How often the pruner thread shall prune the store (sec), if enabled.

interpCfg :: Config

Parameters of timeseries query interpretation.

data TimeseriesHandle

The constructor and projections are not exported. Use the API methods below for working with timeseries.

type QueryId = Word64

create :: forall s. Store s Double => Trace IO TimeseriesTrace -> Maybe TimeseriesConfig -> IO TimeseriesHandle

modifyConfig :: TimeseriesHandle -> (TimeseriesConfig -> Maybe TimeseriesConfig) -> IO ()

Reconfigure the store. The new parameters are applied immediately. Wakes up the pruner thread as well. If the supplied config is empty, reconfigures back to the default config.

readConfig :: TimeseriesHandle -> IO TimeseriesConfig

Read a snapshot of the configuration.

writeConfig :: TimeseriesHandle -> Maybe TimeseriesConfig -> IO ()

A specialised version of modifyConfig where the value of the new config doesn't depend on the old one.

insert :: TimeseriesHandle -> Text -> Text -> Timestamp -> [(MetricIdentifier, Double)] -> IO ()

Insert a batch on metric data into the store at the given timestamp.

execute :: TimeseriesHandle -> Timestamp -> Text -> IO (Either ExecutionError Value)

Execute a query on the store at the given timestamp (ms).

prune :: TimeseriesHandle -> IO ()

Prune the store of the entries past the configured retention period (relative to "now"). The pruner thread uses that function internally, but one can still invoke the pruning manually at will.