Cardano.Timeseries.Component.Types

Documentation

type QueryId = Word64

data TimeseriesConfig

Constructors

TimeseriesConfig

Fields

retentionMillis :: Word64

How long the store entries are retained for (ms).

pruningPeriodSec :: Maybe Word64

How often the pruner thread shall prune the store (sec), if enabled.

interpCfg :: Config

Parameters of timeseries query interpretation.