Cardano.Timeseries.Domain.Timeseries

Documentation

data Timeseries a

A collection of datapoints sharing a series.

Constructors

Timeseries

Fields

labels :: SeriesIdentifier

dat :: [(Timestamp, a)]

type TimeseriesVector a = [Timeseries a]

Every two elements in the list must have distinct series identifiers (set of labels), | i.e. the series in the list must be distinct.

transpose :: [InstantVector a] -> TimeseriesVector a

Given a list of range vectors, forms up a timeseries vector. | This operation is, in some sense, transposition: | | ⎴ ⎴ ⎴ ⎴ | series1: ... ◯ ◯ ... | series2: ... ◯ ◯ ... | series3: ... ◯ ◯ ◯ ... | ... ... | ⎵ ⎵ ⎵ ⎵ | -------------------------------> t | t₀ t₁ t₂ t₃ | | =====> | | | | series1: [ ... ◯ ◯ ... ] | series2: [ ... ◯ ◯ ... ] | series3: [ ... ◯ ◯ ◯ ... ] | ... ... | | ----------------------------------------> t t₀ t₁ t₂ t₃

toVector :: Timeseries Double -> Vector Double

oldest :: Timeseries a -> Maybe (Instant a)

newest :: Timeseries a -> Maybe (Instant a)

eachOldest :: TimeseriesVector a -> Maybe [Instant a]

eachNewest :: TimeseriesVector a -> Maybe [Instant a]

superseries :: Set Label -> SeriesIdentifier -> SeriesIdentifier