Cardano.Timeseries.Elab.Resolve

Documentation

resolveTy :: Defs -> Ty -> Ty

Computes the head-normal form of Ty w.r.t. hole resolution (i.e. unfolds holes recursively up to the head expression).

resolveBinding :: Defs -> Binding -> Binding

Computes the head-normal form of Binding w.r.t. hole resolution (i.e. unfolds holes recursively up to the head expression in type of the binding).

resolveContext :: Defs -> Context -> Context

Computes the head-normal form of Context w.r.t. hole resolution (i.e. unfolds holes recursively up to the head expression in every type of the context).

resolveExpr' :: Defs -> Expr -> Expr

Computes the normal form of Expr w.r.t. hole resolution (i.e. resolves *all* holes in the expression).