Cardano.Timeseries.Elab.Typing

Documentation

data Ty

Typing of a query expression.

Constructors

InstantVector Ty

RangeVector Ty

Scalar

Text

Bool

Pair Ty Ty

Timestamp

Duration

Fun Ty Ty

Hole HoleIdentifier

data Binding

A context entry of a typing context.

Constructors

LetBinding Identifier Expr Ty

LambdaBinding Identifier Ty

type Context = Seq Binding

Γ A typing context of a query expression.

identifier :: Binding -> Identifier

data Def

(? type) | (? ≔ T type) | (? : T) | (? ≔ t : T) Definition of a type- or expression- level hole.

Constructors

TyHoleDecl

TyHoleInst Ty

ExprHoleDecl Ty

ExprHoleInst Expr Ty

type Defs = Map HoleIdentifier Def

Σ A collection of hole definitions Def indexed by HoleIdentifier.

instantiateTy :: HoleIdentifier -> Ty -> Defs -> Defs

Assumes that the given Defs contains a TyHoleDecl of the given HoleIdentifier.

instantiateExpr :: HoleIdentifier -> Expr -> Defs -> Defs

Assumes that the given Defs contains a ExprHoleDecl of the given HoleIdentifier. Types of the hole and the provided Expr must be compatible.

prettyTy :: TyPrec -> Ty -> Text

data TyPrec

Constructors

Loose

FunCodomain

FunDomain

Tight

ty :: Binding -> Ty