Cardano.Timeseries.Elab.Unify

Documentation

data UnificationProblem

A = B type An equation between two query types containing holes. Unification is an algorithm of finding unique solutions to such equations.

Constructors

UnificationProblem

Fields

loc :: Loc

Source location that induces the unification problem.

lhs :: Ty

rhs :: Ty

data St

Constructors

St

Fields

defs :: Defs

type UnifyM a = ExceptT UnifyError (State St) a

unify :: Loc -> Ty -> Ty -> UnifyM [UnificationProblem]

solve :: [UnificationProblem] -> UnifyM ()

Solve the list of unification problems, instantiating holes in the process. If a problem doesn't have a (unique) solution, throw an error.