Cardano.Timeseries.Interp

Documentation

interp :: Store s Double => Config -> s -> Map Identifier Value -> Expr -> Timestamp -> InterpM Value

Given a metric store, an assignment of values to local variables, a query expression and a timestamp "now", interpret the Expr into a Value.