Cardano.Timeseries.Store

Documentation

class Store s a | s -> a where

A type-class witnessing that s is a metric-store of a.

Methods

insert :: s -> MetricIdentifier -> Instant a -> s

Insert an instant into the store under a metric name.

truncate :: s -> Timestamp -> s

Delete all entries older than the specified Timestamp from the store

evaluate :: s -> MetricIdentifier -> Timestamp -> InstantVector a

Compute a point vector of type a such that the timestamp of every point in the vector | lies within the staleness window of the target timestamp and is the most recent of all | points in the store sharing a series.

earliest :: s -> MetricIdentifier -> Maybe Timestamp

Find the earliest occurrence of the metric in the store, if any.

latest :: s -> MetricIdentifier -> Maybe Timestamp

Find the latest occurrence of the metric in the store, if any.

new :: s

An empty store.

metrics :: s -> Set MetricIdentifier

count :: s -> Int

Total number of (metric, labels, timestamp, value) tuples.

stalenessConstant :: Word64

Milliseconds