Cardano.Timeseries.API

Documentation

data ExecutionError where

Constructors

ParsingErrorWhileExecuting

Fields

:: { message :: Text

} -> ExecutionError

ElabErrorWhileExecuting

Fields

:: { message :: Text

} -> ExecutionError

InterpErrorWhileExecuting

Fields

:: { message :: Text

} -> ExecutionError

execute :: Store s Double => s -> Config -> Timestamp -> Text -> Either ExecutionError Value

A pure interface to the composite of parsing, elaboration and interpretation of timeseries queries.

type Parser = Parsec Void Text

type Tree a = Map MetricIdentifier (Map Timestamp [Point a])

type Flat a = [Point a]

data Value where

A model of values that queries interpret into.

Constructors

Scalar :: Double -> Value
A scalar.

RangeVector :: TimeseriesVector Value -> Value
A range vector.

InstantVector :: InstantVector Value -> Value
An instant vector.

Pair :: Value -> Value -> Value
A pair.

Truth :: Value
Truth.

Falsity :: Value
Falsity.

Duration :: Word64 -> Value
Duration (milliseconds)

Timestamp :: Word64 -> Value
Timestamp (milliseconds since epoch)

Function :: FunctionValue -> Value
Function

Text :: Text -> Value
Text

data Config

Constructors

Config

Fields

defaultRangeSamplingRateMillis :: Word64

class AsText a where

For the purpose of pretty-printing. Result may include linebreaks.

Methods

asText :: a -> Text

type MetricIdentifier = Text

Each series in the (metric) store can be identified by a metric name.

type Timestamp = Word64

Milliseconds since epoch.

newtype InterpError

Constructors

InterpError

Fields

message :: Text

type InterpM a = ExceptT InterpError (State Int) a

class Store s a | s -> a where

A type-class witnessing that s is a metric-store of a.

Methods

insert :: s -> MetricIdentifier -> Instant a -> s

Insert an instant into the store under a metric name.

truncate :: s -> Timestamp -> s

Delete all entries older than the specified Timestamp from the store

evaluate :: s -> MetricIdentifier -> Timestamp -> InstantVector a

Compute a point vector of type a such that the timestamp of every point in the vector | lies within the staleness window of the target timestamp and is the most recent of all | points in the store sharing a series.

earliest :: s -> MetricIdentifier -> Maybe Timestamp

Find the earliest occurrence of the metric in the store, if any.

latest :: s -> MetricIdentifier -> Maybe Timestamp

Find the latest occurrence of the metric in the store, if any.

new :: s

An empty store.

metrics :: s -> Set MetricIdentifier

count :: s -> Int

Total number of (metric, labels, timestamp, value) tuples.

double :: Parser Double

point :: Show a => Parser a -> Parser (Point a)

s ::= doublequoted-string s [s = s, ..., s = s] decimal-nat floating-point

showT :: Show a => a -> Text

Convenience function for a Show instance to be converted to text immediately

elab :: Expr -> ElabM Expr

Given a surface query Expr elaborate it into a regular "semantic" query Expr.

initialSt :: St

expr :: Parser Expr

interp :: Store s Double => Config -> s -> Map Identifier Value -> Expr -> Timestamp -> InterpM Value

Given a metric store, an assignment of values to local variables, a query expression and a timestamp "now", interpret the Expr into a Value.

fromFlat :: Flat a -> Tree a